Update quadcopter-maze.md
Updated to new algorithm -> simple DFS (I think simple DFS is better anyway because Wilson's algorithm is harder to implement and Wilson's algorithm is only "better" [or makes the maze look more natural] if the maze is BIG)

Deleted info section. I don't understand it myself and don't think it's applicable to the newly updated maze game

Deleted the old 2D/3D option. Why have 2D when your 3D is perfect and works? Also I would need to code for that which would be a pain in the a**.